xen/console: Use request_platform_preferred_console() to prefer console
There is a new API which allows to distinguish three priorities
for adding preferred consoles.

The term "platform" is used for consoles preferred by HW or SW platform.
They might be defined by, for example, SPCR or device tree. Or they
might be hardcoded in arch-specific code.

The priorities are defined by enum console_pref_type. The meaning is
the following:

  + CON_PREF_PLATFORM_PROPOSAL: Consoles proposed by the platform
	might be overridden by the command line. Otherwise, they will
	be enabled as the default console(s).

  + CON_PREF_PLATFORM_REQUEST: Console requested by the platform
	will be enabled in parallel with consoles defined on
	the command line. Anyway, they are enabled by default.

  + CON_PREF_CMDLINE: Consoles defined on the command line might
	override kernel default or consoles preferred by the platform.

xen_add_preferred_consoles() added the preferred consoles without
checking the command line. It is the 2nd case.

Signed-off-by: Petr Mladek <pmladek@suse.com>