of/console: Use propose_platform_preferred_console() to prefer console
There is a new API which allows to distinguish three priorities
for adding preferred consoles.

The term "platform" is used for consoles preferred by HW or SW platform.
They might be defined by, for example, SPCR or device tree. Or they
might be hard-coded in arch-specific code.

The priorities are defined by enum console_pref_type. The meaning is
the following:

  + CON_PREF_PLATFORM_PROPOSAL: Consoles proposed by the platform
	might be overridden by the command line. Otherwise, they will
	be enabled as the default console(s). They will also
	be used when "console=platform" is used on the command line.

  + CON_PREF_PLATFORM_REQUEST: Console requested by the platform
	will be enabled in parallel with consoles defined on
	the command line. Anyway, they are enabled by default.

  + CON_PREF_CMDLINE: Consoles defined on the command line might
	override kernel default or consoles preferred by the platform.

The consoles proposed by the device tree fits into the 1st category.
Add the preferred console as a proposal. The check of the command line
will be done internally during the console registration.

The change allows to maintain the logic inside the printk code.

And more importantly, it allows to enable the proposed consoles by
the new generic "console=platform" option.

FIXME: Test build

Signed-off-by: Petr Mladek <pmladek@suse.com>