tty/console: Use new *_platform_preferred_console() API to prefer console
There is a new API which allows to distinguish three priorities
for adding preferred consoles.

The term "platform" is used for consoles preferred by HW or SW platform.
They might be defined by, for example, SPCR or device tree. Or they
might be hard-coded in arch-specific code.

The priorities are defined by enum console_pref_type. The meaning is
the following:

  + CON_PREF_PLATFORM_PROPOSAL: Consoles proposed by the platform
	might be overridden by the command line. Otherwise, they will
	be enabled as the default console(s). They will also
	be used when "console=platform" is used on the command line.

  + CON_PREF_PLATFORM_REQUEST: Console requested by the platform
	will be enabled in parallel with consoles defined on
	the command line. Anyway, they are enabled by default.

  + CON_PREF_CMDLINE: Consoles defined on the command line might
	override kernel default or consoles preferred by the platform.

Most of the tty console drivers are requested by the platform.
The only exception are "vio" and "sun" drivers where the console
is only proposed and can be overridden via the command line.

The change allows to maintain the logic inside the printk code.

And more importantly, it allows to enable the proposed consoles by
the new generic "console=platform" option.

FIXME: Test build

Signed-off-by: Petr Mladek <pmladek@suse.com>