powerpc/console: Use prefer_console() API
FIXME: Generic description

Use CON_PREF_PLATFORM_DEFAULT instead of the check of the command line.
It allows to maintain the logic on a single place. And more importantly,
the console can be selected by the generic console=platform.

Use CON_PREF_PLATFORM_REQUEST on (FIXME) because it should always
be enabled on this platform.

Signed-off-by: Petr Mladek <pmladek@suse.com>