printk/console/spcr: Better define when console provided by SPCR is enabled
FIXME: Update the commit message to match the behavior.

       Add console=platform option later?
       Would allow to change the behavior in a separate patch.
       But it might actually be better to see this in one place (patch).

There is a new API which allows to distinguish three priorities
for adding preferred consoles.

The term "platform" is used for consoles preferred by HW or SW platform.
They might be defined by, for example, SPCR or device tree. Or they
might be hard-coded in arch-specific code.

The priorities are defined by enum console_pref_type. The meaning is
the following:

  + CON_PREF_PLATFORM_PROPOSAL: Consoles proposed by the platform
	might be overridden by the command line. Otherwise, they will
	be enabled as the default console(s).

  + CON_PREF_PLATFORM_REQUEST: Console requested by the platform
	will be enabled in parallel with consoles defined on
	the command line. Anyway, they are enabled by default.

  + CON_PREF_CMDLINE: Consoles defined on the command line might
	override kernel default or consoles preferred by the platform.

The situation with SPCR is a bit more complicated. The console proposed
by SPCR is:

  + basically ignored on "x86" and "loongarch" architectures.

  + requested on "arm" architecture unless "acpi=nospcr" is defined on
    the command line.

  + requested on "riscv" architecture unless acpi is disabled

By other words, the proposed console is either requested or ignored.
The variant is passed via the @enable_console parameter passed
to acpi_parse_spcr(). Use the API requesting the console there.

Signed-off-by: Petr Mladek <pmladek@suse.com>